requirements.txt
# method 1 - using wifi_qrcode_generator library
wifi_qrcode_generator

# method 2 - using segno
segno              # qr_code_object code library
qrcode-artistic    # helpers to make tweaks to final image
matplotlib         # fix pycharm complaints

# method 3 - using qrcode
qrcode[pil]==8.2
# below repos required for png output with `qrcode` lib
git+https://github.com/ojii/pymaging.git#egg=pymaging
git+https://github.com/ojii/pymaging-png.git#egg=pymaging-png

pillow==11.3.0
plotly    # via pycharm; just get latest version